Cleanup manager scripts
