+++ includes/manager/src/register-fields-ui.js
@@ -1,0 +1,8 @@
+import { registerPlugin } from '@wordpress/plugins';
+import { FieldsUI } from './components/fields-ui';
+
+export const registerFieldsUI = () => {
+	registerPlugin( 'create-content-model-fields-ui', {
+		render: FieldsUI,
+	} );
+};
