+++ webpack.config.js
@@ -8,9 +8,6 @@ const includesDir = path.resolve( process.cwd(), 'includes' );
 
 const entries = glob
 	.sync( '*/*.js', { cwd: includesDir } )
-	.filter( ( entry ) => {
-		return ! entry.split( '/' )[ 1 ].startsWith( '_' );
-	} )
 	.reduce( ( acc, entry ) => {
 		const [ folder, name ] = entry.split( '/' );
